src/services/rag_techniques/examples/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-techniques"
version = "0.1.0"
description = "A comprehensive toolkit for building advanced RAG systems"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "RAG Techniques Contributors", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.0",
    "pypdf>=3.0.0",
    "PyMuPDF>=1.23.0",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
evaluation = [
    "deepeval>=3.0.0",
]
cloud = [
    "langchain-cohere",
    "boto3>=1.28.0",
]

[project.urls]
Homepage = "https://github.com/NirDiamant/RAG_Techniques"
Documentation = "https://github.com/NirDiamant/RAG_Techniques"
Repository = "https://github.com/NirDiamant/RAG_Techniques"
Issues = "https://github.com/NirDiamant/RAG_Techniques/issues"

[project.scripts]
rag-simple = "rag_techniques.cli:simple_rag_cli"

[tool.setuptools]
packages = ["rag_techniques", "rag_techniques.core", "rag_techniques.techniques", "rag_techniques.utils", "rag_techniques.evaluation"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
